nlp-service/requirements.txt
fastapi>=0.100.0
uvicorn[standard]>=0.23.0
langchain>=0.1.0
langchain-community>=0.0.20
langchain-openai>=0.0.5
langchain-google-genai>=0.0.5
pymongo>=4.6.0
pydantic>=2.0.0,<3.0.0
pydantic-settings>=2.0.0
python-dotenv>=1.0.0
pytest>=7.4.0
pytest-asyncio>=0.21.0
httpx>=0.24.0
nltk>=3.8.1
google-genai>=1.0.0
